Test contracts
As i have never done testing smart contracts with Solidity,  I am just getting the hang of truffle testing by practising  on a simple smart contract : Migrations.sol the default one when a project is created.  So here is the contract Migrations.sol. Created a simple function called sum and added one function from our kol contract : deleteBlock. In the test file that i am comitting in the next commit , there are three tests : one that asserts that every contract has a owner that is the msg.sender ; one that tests the sum function and the last one tests the delete block function of the kol contract.